[win10] powermanagement: move implementation to platform folder
