+++ cmake/treedata/windowsstore/subdirs.txt
@@ -1,15 +1,15 @@
-xbmc/platform/win10             platform/win10
-xbmc/platform/win10/filesystem  platform/win10/filesystem
-xbmc/platform/win10/network     platform/win10/network
-xbmc/platform/win10/peripherals platform/win10/peripherals
-xbmc/platform/win10/storage     platfrom/win10/storage
-xbmc/platform/win32/filesystem  platform/win32/filesystem
-xbmc/input/touch                input/touch
-xbmc/input/touch/generic        input/touch/generic
-xbmc/network/mdns               network/mdns
-xbmc/powermanagement/win10      powermanagement/win10
-xbmc/utils/win32                utils/win32
-xbmc/rendering/dx               rendering/dx
-xbmc/threads/platform/win       threads/win
-xbmc/windowing/win10            windowing/win10
+xbmc/platform/win10                    platform/win10
+xbmc/platform/win10/filesystem         platform/win10/filesystem
+xbmc/platform/win10/network            platform/win10/network
+xbmc/platform/win10/peripherals        platform/win10/peripherals
+xbmc/platform/win10/powermanagement    platfrom/win10/powermanagement
+xbmc/platform/win10/storage            platfrom/win10/storage
+xbmc/platform/win32/filesystem         platform/win32/filesystem
+xbmc/input/touch                       input/touch
+xbmc/input/touch/generic               input/touch/generic
+xbmc/network/mdns                      network/mdns
+xbmc/utils/win32                       utils/win32
+xbmc/rendering/dx                      rendering/dx
+xbmc/threads/platform/win              threads/win
+xbmc/windowing/win10                   windowing/win10
 xbmc/cores/VideoPlayer/Process/windows cores/VideoPlayer/Process/windows
