+++ xbmc/platform/win10/powermanagement/CMakeLists.txt
@@ -3,4 +3,4 @@ set(SOURCES Win10PowerSyscall.cpp
 
 set(HEADERS Win10PowerSyscall.h)
 
-core_add_library(powermanagement_win10)
+core_add_library(platform_win10_powermanagement)
