+++ xbmc/platform/win10/powermanagement/Win10PowerSyscall.cpp
