+++ xbmc/platform/win10/powermanagement/WinIdleTimer.cpp
