+++ xbmc/windowing/win10/WinSystemWin10.cpp
@@ -27,8 +27,8 @@
 #include "messaging/ApplicationMessenger.h"
 #include "platform/win10/AsyncHelpers.h"
 #include "platform/win10/input/RemoteControlXbox.h"
+#include "platform/win10/powermanagement/Win10PowerSyscall.h"
 #include "platform/win32/CharsetConverter.h"
-#include "powermanagement/win10/Win10PowerSyscall.h"
 #include "rendering/dx/DirectXHelper.h"
 #include "rendering/dx/RenderContext.h"
 #include "ServiceBroker.h"
